pyproject.toml
[tool.poetry]
name = "django-wallet"
version = "0.0.1"
description = "A django application to show transactional outbox pattern for simple wallets"
license = "MIT"
authors = ["Aleksandr Bogut <chaotism@mail.ru>"]
readme = "README.md"
homepage = "https://github.com/chaotism/django-wallet"
repository = "https://github.com/chaotism/django-wallet"
documentation = "https://github.com/chaotism/django-wallet"
keywords = ["transactional outbox patterns", "wallet", "microservices"]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Framework :: Django :: 5.0',
    'Framework :: Django :: 5.1',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.13',
    'Topic :: Internet :: WWW/HTTP',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
]
packages = [
    { include = "django_wallet" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
asgiref = "^3.8.1"
django = "^5.1.2"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
djangorestframework-jsonapi = "^7.0.2"
drf-yasg = "^1.21.8"
inflection = "^0.5.1"
mysqlclient = "^2.2.5"
packaging = "^24.1"
pytz = "^2024.2"
PyYAML = "^6.0.2"
sqlparse = "^0.5.1"
uritemplate = "^4.1.1"
load-dotenv = "^0.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
coverage = "*"
pre-commit = "^3.8.0"

[tool.black]
line-length = 120
target_version = ["py310"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
[tool.isort]
force_single_line = true
line_length = 120
py_version = 312
use_parentheses = true
multi_line_output = 3
include_trailing_comma = true
lines_between_types = 1
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_wallet.settings"

# flake8-mypy expects the two following for sensible formatting
show_column_numbers = true
show_error_context = false

# do not follow imports (except for ones found in typeshed)
follow_imports = "skip"

# suppress errors about unsatisfied imports
ignore_missing_imports = true

# allow untyped calls as a consequence of the options above
disallow_untyped_calls = false

# allow returning Any as a consequence of the options above
warn_return_any = false

# treat Optional per PEP 484
strict_optional = true

# ensure all execution paths are returning
warn_no_return = true

# lint-style cleanliness for typing needs to be disabled; returns more errors
# than the full run.
warn_redundant_casts = false
warn_unused_ignores = false

# The following are off by default since they're too noisy.
# Flip them on if you feel adventurous.
disallow_untyped_defs = false
check_untyped_defs = false

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
